Changing selectors to be an array; adding files type; adding prefix option
Selectors is now an array to support multiple matchers of the same name,
but different file path. Files can now be specified as an array to only
match specific file paths without crazy regex statements. This change
also adds support for adding a prefix to the title of the resource.